Made Timeout an interface.
