+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/timeout/Timeout.java
@@ -1,71 +1,46 @@
 package de.uka.ipd.sdq.beagle.core.timeout;
 
-import org.apache.commons.lang3.Validate;
-
-import java.util.HashSet;
-import java.util.Set;
-
 /**
- * Decides when a timeout is reached. This abstract class has implementing classes which
- * do so depending on different criteria.
+ * Policy deciding when to stop the execution of a program. Clients can report progress
+ * through {@link #reportOneStepProgress()}. This information may be used by the policy to
+ * make a decision. Once the policy decides to end the execution, it will call all
+ * handlers registered through {@link #registerCallback(Runnable)}. This will happen
+ * asynchronously.
  *
+ * <p>A timeout must be initialised through {@link #init()} exactly once before it is
+ * used.
+ *
+ * @author Joshua Gleitze
  * @author Christoph Michelbach
  */
-public abstract class Timeout {
+public interface Timeout {
 
 	/**
-	 * Indicates whether this object has been initialised.
-	 */
-	protected boolean initialised;
-
-	/**
-	 * The millisecond timestamp when this Timeout object has been initialised.
-	 */
-	protected long startingTime;
-
-	/**
-	 * A set of callbacks which will be called once the timeout is reached.
-	 */
-	protected Set<Runnable> callbacks = new HashSet<>();
-
-	/**
-	 * Determines whether the timeout is reached.
+	 * Determines whether the timeout is reached. Once this method has returned
+	 * {@code true}, it will return {@code true} on all further ivocations.
 	 *
 	 * @return {@code true} if the timeout is reached; {@code false} otherwise.
 	 */
-	public abstract boolean isReached();
+	boolean isReached();
 
 	/**
-	 * Call this method after each step of the work the timeout is about.
+	 * Reports one step of progress to the policy. What one step of progress is is to be
+	 * defined by the client. The timeout may use this information to make its decision.
 	 */
-	public abstract void reportOneStepProgress();
+	void reportOneStepProgress();
 
 	/**
-	 * Initialises the timeout object. Sets the starting time to the current time. Must be
-	 * called before any different method is called and must only be called once.
+	 * Initialises the timeout object. This method must be called exactly once before any
+	 * other method is called.
 	 */
-	public final void init() {
-		Validate.validState(!this.initialised);
-
-		this.startingTime = System.currentTimeMillis();
-		this.initialised = true;
-
-		this.implementationInit();
-	}
-
-	/**
-	 * For init methots of implemanting classes. Runs after {@link #init()}.
-	 */
-	public abstract void implementationInit();
+	void init();
 
 	/**
 	 * Registers {@code callback} to be run once the timeout is reached.
 	 *
 	 * @param callback A {@link Runnable} object.
 	 */
-	public void registerCallback(final Runnable callback) {
-		this.callbacks.add(callback);
-	}
+	void registerCallback(final Runnable callback);
 
 	/**
 	 * Removes {@code callback} from the set of callbacks to be run once the timeout is
@@ -73,7 +48,5 @@ public abstract class Timeout {
 	 *
 	 * @param callback A {@link Runnable} object.
 	 */
-	public void unregisterCallback(final Runnable callback) {
-		this.callbacks.remove(callback);
-	}
+	void unregisterCallback(final Runnable callback);
 }
