+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/timeout/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Provides the interface {@link Timeout} and implementations of it. The implementing
+ * Provides the {@link Timeout} interface and implementations of it. The implementing
  * classes decide when a timeout is reached.
  */
 package de.uka.ipd.sdq.beagle.core.timeout;
